+++ sdk/core/azure-core/src/datetime.cpp
@@ -6,6 +6,7 @@
 #include <cstdio>
 #include <cstring>
 #include <limits>
+#include <stdexcept>
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
@@ -38,7 +39,7 @@ DateTime DateTime::Now()
   }
 
   int64_t result = WindowsToPosixOffsetSeconds + time.tv_sec;
-  result *= Duration::IntervalsOf100nsPerSecond; // convert to 10e-7
+  result *= DateTime::Duration::period::den; // convert to 10e-7
   result += time.tv_usec * 10; // convert and add microseconds, 10e-6 to 10e-7
 
   return DateTime(Duration(result));
@@ -135,15 +136,14 @@ constexpr char const monthNames[] = "Jan\0Feb\0Mar\0Apr\0May\0Jun\0Jul\0Aug\0Sep
 std::string DateTime::GetString(DateFormat format, TimeFractionFormat fractionFormat) const
 {
   static constexpr auto const EndOfYear9999 = 2650467743999999999LL;
-  if (m_since1601.m_100nsIntervals > EndOfYear9999)
+  if (m_since1601.count() > EndOfYear9999)
   {
-    throw DateTimeException("The requested year exceeds the year 9999.");
+    throw std::length_error("The requested year exceeds the year 9999.");
   }
 
-  int64_t const epochAdjusted = m_since1601.m_100nsIntervals;
-  int64_t const secondsSince1601
-      = epochAdjusted / Duration::IntervalsOf100nsPerSecond; // convert to seconds
-  int const fracSec = static_cast<int>(epochAdjusted % Duration::IntervalsOf100nsPerSecond);
+  int64_t const epochAdjusted = m_since1601.count();
+  int64_t const secondsSince1601 = epochAdjusted / Duration::period::den; // convert to seconds
+  int const fracSec = static_cast<int>(epochAdjusted % Duration::period::den);
 
   auto const yearData = ComputeYear(secondsSince1601);
   int const year = yearData.Year + MinYear;
@@ -197,7 +197,7 @@ std::string DateTime::GetString(DateFormat format, TimeFractionFormat fractionFo
       outCursor += 4;
       return std::string(outBuffer, outCursor);
 
-    case DateFormat::Iso8601:
+    case DateFormat::Rfc3339:
 #ifdef _MSC_VER
       sprintf_s(
 #else
@@ -249,7 +249,7 @@ std::string DateTime::GetString(DateFormat format, TimeFractionFormat fractionFo
       return std::string(outBuffer, outCursor);
 
     default:
-      throw DateTimeException("Unrecognized date format.");
+      throw std::invalid_argument("Unrecognized date format.");
   }
 }
 
@@ -405,9 +405,6 @@ zone        =  "UT"  / "GMT"                ; Universal Time
 
 DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
 {
-  auto const zeroDuration = Duration();
-  DateTime result(zeroDuration);
-
   int64_t secondsSince1601 = 0;
   uint64_t fracSec = 0;
 
@@ -438,12 +435,12 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     }
     else
     {
-      throw DateTimeException("Error parsing DateTime: Day of the month.");
+      throw std::invalid_argument("Error parsing DateTime: Day of the month.");
     }
 
     if (monthDay == 0)
     {
-      throw DateTimeException("Error parsing DateTime: Invalid month day number (0).");
+      throw std::invalid_argument("Error parsing DateTime: Invalid month day number (0).");
     }
 
     int month = 0;
@@ -457,13 +454,13 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
       ++month;
       if (month == 12)
       {
-        throw DateTimeException("Error parsing DateTime: Month number.");
+        throw std::invalid_argument("Error parsing DateTime: Month number.");
       }
     }
 
     if (str[3] != ' ')
     {
-      throw DateTimeException("Error parsing DateTime.");
+      throw std::invalid_argument("Error parsing DateTime.");
     }
 
     str += 4; // parsed month
@@ -471,19 +468,19 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     if (!IsDigit(str[0]) || !IsDigit(str[1]) || !IsDigit(str[2]) || !IsDigit(str[3])
         || str[4] != ' ')
     {
-      throw DateTimeException("Error parsing DateTime: Year.");
+      throw std::invalid_argument("Error parsing DateTime: Year.");
     }
 
     int year = (str[0] - '0') * 1000 + (str[1] - '0') * 100 + (str[2] - '0') * 10 + (str[3] - '0');
     if (year < MinYear)
     {
-      throw DateTimeException("Error parsing DateTime: Year is less than 1601.");
+      throw std::invalid_argument("Error parsing DateTime: Year is less than 1601.");
     }
 
     // days in month validity check
     if (!ValidateDay(monthDay, month, year))
     {
-      throw DateTimeException("Error parsing DateTime: Invalid day in month.");
+      throw std::invalid_argument("Error parsing DateTime: Invalid day in month.");
     }
 
     str += 5; // parsed year
@@ -492,13 +489,13 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     if (!IsDigit<2>(str[0]) || !IsDigit(str[1]) || str[2] != ':' || !IsDigit<5>(str[3])
         || !IsDigit(str[4]))
     {
-      throw DateTimeException("Error parsing DateTime: Hour and minutes.");
+      throw std::invalid_argument("Error parsing DateTime: Hour and minutes.");
     }
 
     int const hour = StringToDoubleDigitInt(str);
     if (hour > 23)
     {
-      throw DateTimeException("Error parsing DateTime: hour > 23.");
+      throw std::invalid_argument("Error parsing DateTime: hour > 23.");
     }
     str += 3; // parsed hour
 
@@ -510,7 +507,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     {
       if (!IsDigit<6>(str[1]) || !IsDigit(str[2]) || str[3] != ' ')
       {
-        throw DateTimeException("Error parsing DateTime.");
+        throw std::invalid_argument("Error parsing DateTime.");
       }
 
       sec = StringToDoubleDigitInt(str + 1);
@@ -522,12 +519,12 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     }
     else
     {
-      throw DateTimeException("Error parsing DateTime.");
+      throw std::invalid_argument("Error parsing DateTime.");
     }
 
     if (sec > 60)
     { // 60 to allow leap seconds
-      throw DateTimeException("Error parsing DateTime: Seconds > 60.");
+      throw std::invalid_argument("Error parsing DateTime: Seconds > 60.");
     }
 
     year -= MinYear;
@@ -538,7 +535,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
       int const actualWeekday = (daysSince1601 + 1) % 7;
       if (parsedWeekday != actualWeekday)
       {
-        throw DateTimeException("Error parsing DateTime: Weekday.");
+        throw std::invalid_argument("Error parsing DateTime: Weekday.");
       }
     }
 
@@ -583,7 +580,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
       }
       else
       {
-        throw DateTimeException("Error parsing DateTime: Time zone.");
+        throw std::invalid_argument("Error parsing DateTime: Time zone.");
       }
 
       secondsSince1601
@@ -591,23 +588,23 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
 
       if (secondsSince1601 < 0)
       {
-        throw DateTimeException(
+        throw std::invalid_argument(
             "Error parsing DateTime: year is < 1601 after time zone adjustments.");
       }
     }
   }
-  else if (format == DateFormat::Iso8601)
+  else if (format == DateFormat::Rfc3339)
   {
     // parse year
     if (!IsDigit(str[0]) || !IsDigit(str[1]) || !IsDigit(str[2]) || !IsDigit(str[3]))
     {
-      throw DateTimeException("Error parsing DateTime: Year.");
+      throw std::invalid_argument("Error parsing DateTime: Year.");
     }
 
     int year = (str[0] - '0') * 1000 + (str[1] - '0') * 100 + (str[2] - '0') * 10 + (str[3] - '0');
     if (year < MinYear)
     {
-      throw DateTimeException("Error parsing DateTime: Year < 1601.");
+      throw std::invalid_argument("Error parsing DateTime: Year < 1601.");
     }
 
     str += 4;
@@ -619,13 +616,13 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     // parse month
     if (!IsDigit<1>(str[0]) || !IsDigit(str[1]))
     {
-      throw DateTimeException("Error parsing DateTime: Month number.");
+      throw std::invalid_argument("Error parsing DateTime: Month number.");
     }
 
     int month = StringToDoubleDigitInt(str);
     if (month < 1 || month > 12)
     {
-      throw DateTimeException("Error parsing DateTime: INvalid month number.");
+      throw std::invalid_argument("Error parsing DateTime: Invalid month number.");
     }
 
     month -= 1;
@@ -639,13 +636,13 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     // parse day
     if (!IsDigit<3>(str[0]) || !IsDigit(str[1]))
     {
-      throw DateTimeException("Error parsing DateTime: Day.");
+      throw std::invalid_argument("Error parsing DateTime: Day.");
     }
 
     int monthDay = StringToDoubleDigitInt(str);
     if (!ValidateDay(monthDay, month, year))
     {
-      throw DateTimeException("Error parsing DateTime: Day of the month.");
+      throw std::invalid_argument("Error parsing DateTime: Day of the month.");
     }
 
     int const yearDay = GetYearDay(month, monthDay, year);
@@ -658,9 +655,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     {
       // No time
       secondsSince1601 = static_cast<int64_t>(daysSince1601) * SecondsInDay;
-      result.m_since1601.m_100nsIntervals
-          = static_cast<int64_t>(secondsSince1601 * Duration::IntervalsOf100nsPerSecond + fracSec);
-      return result;
+      return DateTime(std::chrono::seconds(secondsSince1601) + DateTime::Duration(fracSec));
     }
 
     ++str; // skip 'T'
@@ -668,14 +663,14 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     // parse hour
     if (!IsDigit<2>(str[0]) || !IsDigit(str[1]))
     {
-      throw DateTimeException("Error parsing DateTime: Hour.");
+      throw std::invalid_argument("Error parsing DateTime: Hour.");
     }
 
     int const hour = StringToDoubleDigitInt(str);
     str += 2;
     if (hour > 23)
     {
-      throw DateTimeException("Error parsing DateTime: Invalid hour number.");
+      throw std::invalid_argument("Error parsing DateTime: Invalid hour number.");
     }
 
     if (*str == ':')
@@ -686,7 +681,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     // parse minute
     if (!IsDigit<5>(str[0]) || !IsDigit(str[1]))
     {
-      throw DateTimeException("Error parsing DateTime: Minute.");
+      throw std::invalid_argument("Error parsing DateTime: Minute.");
     }
 
     int const minute = StringToDoubleDigitInt(str);
@@ -703,14 +698,14 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
     // parse seconds
     if (!IsDigit<6>(str[0]) || !IsDigit(str[1]))
     {
-      throw DateTimeException("Error parsing DateTime: Seconds.");
+      throw std::invalid_argument("Error parsing DateTime: Seconds.");
     }
 
     int const sec = StringToDoubleDigitInt(str);
     // We allow 60 to account for leap seconds
     if (sec > 60)
     {
-      throw DateTimeException("Error parsing DateTime: Invalid seconds number.");
+      throw std::invalid_argument("Error parsing DateTime: Invalid seconds number.");
     }
 
     str += 2;
@@ -764,7 +759,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
       if (!IsDigit<2>(str[1]) || !IsDigit(str[2]) || str[3] != ':' || !IsDigit<5>(str[4])
           || !IsDigit(str[5]))
       {
-        throw DateTimeException("Error parsing DateTime.");
+        throw std::invalid_argument("Error parsing DateTime.");
       }
 
       secondsSince1601 = AdjustTimezone(
@@ -774,7 +769,7 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
           StringToDoubleDigitInt(str + 4));
       if (secondsSince1601 < 0)
       {
-        throw DateTimeException("Error parsing DateTime.");
+        throw std::invalid_argument("Error parsing DateTime.");
       }
     }
     else
@@ -784,12 +779,10 @@ DateTime DateTime::Parse(std::string const& dateString, DateFormat format)
   }
   else
   {
-    throw DateTimeException("Unrecognized date format.");
+    throw std::invalid_argument("Unrecognized date format.");
   }
 
-  result.m_since1601.m_100nsIntervals
-      = static_cast<int64_t>(secondsSince1601 * Duration::IntervalsOf100nsPerSecond + fracSec);
-  return result;
+  return DateTime(std::chrono::seconds(secondsSince1601) + DateTime::Duration(fracSec));
 }
 
 DateTime::DateTime(
@@ -804,39 +797,39 @@ DateTime::DateTime(
   // from this constructor
   if (year > MaxYear || year < MinYear)
   {
-    throw DateTimeException(
+    throw std::invalid_argument(
         year > MaxYear ? "The requested year exceeds the year 9999."
                        : "The requested year is less than the year 1601.");
   }
 
   if (month <= 0 || month > 12)
   {
-    throw DateTimeException("Invalid month value.");
+    throw std::invalid_argument("Invalid month value.");
   }
 
   if (day <= 0 || day > 31)
   {
-    throw DateTimeException("Invalid day value.");
+    throw std::invalid_argument("Invalid day value.");
   }
 
   if (hour < 0 || hour > 23)
   {
-    throw DateTimeException("Invalid hour value.");
+    throw std::invalid_argument("Invalid hour value.");
   }
 
   if (minute < 0 || minute > 60)
   {
-    throw DateTimeException("Invalid minute value.");
+    throw std::invalid_argument("Invalid minute value.");
   }
 
   if (second < 0 || second > 60)
   {
-    throw DateTimeException("Invalid seconds value.");
+    throw std::invalid_argument("Invalid seconds value.");
   }
 
   if (!ValidateDay(day, month - 1, year))
   {
-    throw DateTimeException("Invalid day of the month.");
+    throw std::invalid_argument("Invalid day of the month.");
   }
 
   char outBuffer[38]{}; // Thu, 01 Jan 1970 00:00:00 GMT\0
@@ -864,6 +857,6 @@ DateTime::DateTime(
 
   *outCursor = 'Z';
   ++outCursor;
-  auto const dt = DateTime::Parse(std::string(outBuffer, outCursor), DateFormat::Iso8601);
+  auto const dt = DateTime::Parse(std::string(outBuffer, outCursor), DateFormat::Rfc3339);
   m_since1601 = dt.m_since1601;
 }